refactor(orchestration): one shared SseChunker, drop the two app copies
Candidate #2 (contract-neutral part) from /improve-codebase-architecture.
SseChunker was duplicated internal-static in GaApi and GaChatbot.Api with a
divergent empty-check. Move one public copy to GA.Business.Core.Orchestration
(both hosts already reference it); repoint the 5 call sites; drop the dead
'using Helpers;' left in AgUiChatController. The IChatApplicationService
contract-merge + 'routed' deletion are deliberately NOT done — recorded
codex C-prime 'keep host frozen' guidance (ServiceCollectionExtensions.cs:35).
