+++ Apps/ga-server/GaApi/Controllers/AgUiChatController.cs
@@ -6,7 +6,6 @@ using GA.Business.Core.Orchestration.AgUi;
 using GA.Business.Core.Orchestration.Models;
 using OrchestratorChatRequest = GA.Business.Core.Orchestration.Models.ChatRequest;
 using GA.Business.ML.Agents;
-using Helpers;
 using Services;
 
 /// <summary>
